Separate modules by section #3
