+++ basics/constants.go
@@ -1,0 +1,19 @@
+package basics
+
+import "fmt"
+
+func Constants() {
+	fmt.Println("Learn Go Constants")
+	/* Constants */
+	/*
+		//--------------------------------------------------
+		const Pi = 3.14
+		const (
+			Number = 1234
+			Truth  = false
+			Big    = 1 << 62
+			Small  = Big >> 61
+		)
+		//--------------------------------------------------
+	*/
+}
